[core] Avoid precompute overscaling for TileID
Plus type change for zoom and removed redundant inlines.

+++ src/mbgl/source/source_info.hpp
@@ -17,8 +17,8 @@ class TileID;
 class SourceInfo {
 public:
     std::vector<std::string> tiles;
-    uint16_t minZoom = 0;
-    uint16_t maxZoom = 22;
+    uint8_t minZoom = 0;
+    uint8_t maxZoom = 22;
     std::string attribution;
     LatLng center;
     double zoom = 0;
